update: consolidate corpus loading into _common; generalize verify_drops to verify_corpus
